old_requirements.txt
# altair==4.1.0
# altgraph==0.17
# argon2-cffi==20.1.0
# astor==0.8.1
# async-generator==1.10
# attrs==20.3.0
# autopep8==1.5.4
# backcall==0.2.0
# base58==2.0.1
# bleach==3.2.1
# blinker==1.4
# cachetools==4.2.0
# certifi==2020.6.20
# cffi==1.14.4
# chardet==4.0.0

# colorama==0.4.4
# configparser==5.0.1
# decorator==4.4.2
# defusedxml==0.6.0
# docopt==0.6.2
docx2txt==0.8
# entrypoints==0.3
# fpdf==1.7.2
# future==0.18.2
# gitdb==4.0.5
# GitPython==3.1.11
# idna==2.10
# importlib-metadata==3.3.0
# ipykernel==5.4.2
# ipython==7.19.0
# ipython-genutils==0.2.0
# ipywidgets==7.6.2
# jedi==0.18.0
# Jinja2==2.11.2
# jsonschema==3.2.0
# jupyter-client==6.1.7
# jupyter-core==4.7.0
# jupyterlab-pygments==0.1.2
# jupyterlab-widgets==1.0.0
# MarkupSafe==1.1.1
# minio==7.0.0
# mistune==0.8.4
# nbclient==0.5.1
# nbconvert==6.0.7
# nbformat==5.0.8
# nest-asyncio==1.4.3
# notebook==6.1.6
# numpy==1.19.2
# opencv-python-headless
# packaging==20.8
pandas==1.2.0
# pandocfilters==1.4.3
# parso==0.8.1
# pdfminer.six==20200517
# pdfplumber==0.5.25
# pefile==2019.4.18
# pickleshare==0.7.5
Pillow==8.0.1
# pipreqs==0.4.10
# prometheus-client==0.9.0
# prompt-toolkit==3.0.8
# protobuf==3.14.0
# # pyarrow==2.0.0
# pycodestyle==2.6.0
# pycparser==2.20
# pycryptodome==3.9.9
# pydeck==0.5.0
# Pygments==2.7.3
# pyinstaller==4.0
# pyinstaller-hooks-contrib==2020.9
# pylovepdf==1.3.2
# pyparsing==2.4.7
# pyrsistent==0.17.3
# python-dateutil==2.8.1
# pytz==2020.5
# pywin32-ctypes==0.2.0
# pywinpty==0.5.7
# pyzmq==20.0.0
# requests==2.25.1
# Send2Trash==1.5.0
# six==1.15.0
# smmap==3.0.4
# sortedcontainers==2.3.0
click==7.1.2
streamlit==0.86.0
click==7.1.2
# terminado==0.9.1
# testpath==0.4.4
# toml==0.10.2
# toolz==0.11.1
# tornado==6.1
# traitlets==5.0.5
# typing-extensions==3.7.4.3
# tzlocal==2.1
# urllib3==1.26.2
# validators==0.18.2
# Wand==0.6.5
# watchdog==1.0.2
# wcwidth==0.2.5
# webencodings==0.5.1
# widgetsnbextension==3.5.1
# wincertstore==0.2
# yarg==0.1.9
# zipp==3.4.0
PyPDF2==2.11.0
pytesseract==0.3.10


# freeglut3-dev
# libgtk2.0-dev